user movement mapping
object import, changed xyzhpr to float
added user movement mapping
changed TreeMenu::getDependency to static
added floor and ability to turn it off

+++ src/Object.cpp
@@ -1,7 +1,7 @@
 #include "Object.h"
 #include "selectedObjects.h"
 
-void Object::setHPR(int h, int p, int r)
+void Object::setHPR(float h, float p, float r)
 {
 	arVector3 angles;
 	arEulerAngles* eulers = new arEulerAngles(AR_XYZ, angles);
